# docs/api/openapi-schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kàff GUI API",
        "version": "1.0.0",
        "description": "API REST de gestion de volière colombophile.\n\n## Authentification\n\nL'API utilise JWT (JSON Web Tokens) avec authentification 2FA (TOTP).\n\n### Flux d'authentification\n\n1. **Inscription** : `POST /api/v1/auth/register/` - Retourne un QR code TOTP\n2. **Confirmation TOTP** : `POST /api/v1/auth/2fa/confirm/` - Active le compte\n3. **Connexion** : `POST /api/v1/auth/login/` - Retourne access + refresh tokens\n4. **Utilisation** : Ajouter `Authorization: Bearer <access_token>` dans les headers\n5. **Refresh** : `POST /api/v1/auth/token/refresh/` - Obtenir un nouveau access token\n\n### Sécurité\n\n- Tokens access : 15 minutes de validité\n- Tokens refresh : 7 jours de validité avec rotation\n- Rate limiting : 300 req/min pour utilisateurs authentifiés\n- Protection brute-force : 5 tentatives max, cooldown 5 min\n\nVoir le cahier des charges et les user stories pour le détail métier.",
        "contact": {
            "name": "Serigne Mourtalla Syll",
            "email": "serignemourtallasyll86@gmail.com"
        },
        "license": {
            "name": "Proprietary"
        }
    },
    "paths": {
...
        "/api/v1/pigeons/{id}/declarer-deces/": {
            "post": {
                "operationId": "pigeons_declarer_deces_create",
                "description": "CRUD pigeons + actions de sortie + généalogie.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Pigeon.",
                        "required": true
                    }
                ],
                "tags": [
                    "pigeons"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DecesInputRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/DecesInputRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/DecesInputRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PigeonDetail"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/pigeons/{id}/declarer-perte/": {
            "post": {
                "operationId": "pigeons_declarer_perte_create",
                "description": "CRUD pigeons + actions de sortie + généalogie.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Pigeon.",
                        "required": true
                    }
                ],
                "tags": [
                    "pigeons"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PerteInputRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PerteInputRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PerteInputRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PigeonDetail"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/pigeons/{id}/genealogy/": {
            "get": {
                "operationId": "pigeons_genealogy_retrieve",
                "description": "Arbre généalogique : père, mère, descendants directs (US-REP-03).",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Pigeon.",
                        "required": true
                    }
                ],
                "tags": [
                    "pigeons"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genealogy"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/pigeons/{id}/vendre/": {
            "post": {
                "operationId": "pigeons_vendre_create",
                "description": "CRUD pigeons + actions de sortie + généalogie.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Identifiant"
                        },
                        "description": "A UUID string identifying this Pigeon.",
                        "required": true
                    }
                ],
                "tags": [
                    "pigeons"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VenteInputRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/VenteInputRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/VenteInputRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PigeonDetail"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DecesInputRequest": {
                "type": "object",
                "description": "US-SOR-02 — date obligatoire, cause optionnelle.",
                "properties": {
                    "date_sortie": {
                        "type": "string",
                        "format": "date"
                    },
                    "cause": {
                        "type": "string",
                        "default": "",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "date_sortie"
                ]
            },
...
            "Genealogy": {
                "type": "object",
                "description": "Réponse de l'endpoint /pigeons/<id>/genealogy/ (US-REP-03).",
                "properties": {
                    "pigeon": {
                        "$ref": "#/components/schemas/PigeonNested"
                    },
                    "pere": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PigeonNested"
                            }
                        ],
                        "nullable": true
                    },
                    "mere": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PigeonNested"
                            }
                        ],
                        "nullable": true
                    },
                    "descendants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PigeonNested"
                        }
                    }
                },
                "required": [
                    "descendants",
                    "mere",
                    "pere",
                    "pigeon"
                ]
            },
...
            "PerteInputRequest": {
                "type": "object",
                "description": "US-SOR-03 — date obligatoire, circonstance optionnelle.",
                "properties": {
                    "date_sortie": {
                        "type": "string",
                        "format": "date"
                    },
                    "circonstance": {
                        "type": "string",
                        "default": "",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "date_sortie"
                ]
            },
...
            "PigeonDetail": {
                "type": "object",
                "description": "Fiche complète (US-PIG-02) — parents directs imbriqués, descendants ailleurs.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "title": "Identifiant"
                    },
                    "bague": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "sexe": {
                        "$ref": "#/components/schemas/SexeEnum"
                    },
                    "race": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date"
                    },
                    "age_jours": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "couleur": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "poids": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,3}(?:\\.\\d{0,2})?$",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string"
                    },
                    "statut": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Statut450Enum"
                            }
                        ],
                        "readOnly": true
                    },
                    "pere": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PigeonNested"
                            }
                        ],
                        "readOnly": true
                    },
                    "mere": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/PigeonNested"
                            }
                        ],
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Créé le"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Mis à jour le"
                    }
                },
                "required": [
                    "age_jours",
                    "bague",
                    "created_at",
                    "date_naissance",
                    "id",
                    "mere",
                    "pere",
                    "race",
                    "sexe",
                    "statut",
                    "updated_at"
                ]
            },
...
            "PigeonNested": {
                "type": "object",
                "description": "Représentation minimaliste utilisée en imbriqué (parents, occupant cage…).",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "title": "Identifiant"
                    },
                    "bague": {
                        "type": "string",
                        "readOnly": true
                    },
                    "sexe": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SexeEnum"
                            }
                        ],
                        "readOnly": true
                    },
                    "race": {
                        "type": "string",
                        "readOnly": true
                    },
                    "statut": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Statut450Enum"
                            }
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "bague",
                    "id",
                    "race",
                    "sexe",
                    "statut"
                ]
            },
...
            "SexeEnum": {
                "enum": [
                    "MALE",
                    "FEMALE"
                ],
                "type": "string",
                "description": "* `MALE` - Mâle\n* `FEMALE` - Femelle"
            },
...
            "Statut450Enum": {
                "enum": [
                    "ACTIF",
                    "VENDU",
                    "MORT",
                    "PERDU"
                ],
                "type": "string",
                "description": "* `ACTIF` - Actif\n* `VENDU` - Vendu\n* `MORT` - Mort\n* `PERDU` - Perdu"
            },
...
            "VenteInputRequest": {
                "type": "object",
                "description": "US-SOR-01 — date, prix, acheteur obligatoires.",
                "properties": {
                    "date_sortie": {
                        "type": "string",
                        "format": "date"
                    },
                    "prix": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "acheteur": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                    }
                },
                "required": [
                    "acheteur",
                    "date_sortie",
                    "prix"
                ]
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT access token obtenu via /api/v1/auth/login/"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}